Calculadora finalizada.

+++ Calculadora/src/CalculatorOperation.java
@@ -2,25 +2,25 @@
 enum CalculatorOperation {
 	SOMA("+") {
 		@Override
-		int execute(final int num1, final int num2) {
+		double execute(final double num1, final double num2) {
 			return num1 + num2;
 		}
 	},
 	SUBTRACAO("-") {
 		@Override
-		int execute(final int num1, final int num2) {
+		double execute(final double num1, final double num2) {
 			return num1 - num2;
 		}
 	},
 	MULTIPLICACAO("*") {
 		@Override
-		int execute(final int num1, final int num2) {
+		double execute(final double num1, final double num2) {
 			return num1 * num2;
 		}
 	},
 	DIVISAO("/") {
 		@Override
-		int execute(final int num1, final int num2) {
+		double execute(final double num1, final double num2) {
 			if (num2 == 0) {
 				if (num1 == 0) {
 					throw new ArithmeticException("Operação indefinida");
@@ -42,5 +42,5 @@ enum CalculatorOperation {
 		return value;
 	}
 
-	abstract int execute(int num1, int num2);
+	abstract double execute(double num1, double num2);
 }
